requirements.txt
asgiref==3.4.1
barcode==1.0.1
colorama==0.4.4
Django==3.2.8
Pillow==8.4.0
python-barcode==0.13.1
python-Levenshtein==0.12.0
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
